refactor(auv-driver-macos): move support and capture modules
