+++ crates/auv-driver-macos/src/capture.rs
@@ -1,0 +1,6 @@
+#[doc(hidden)]
+pub mod artifact;
+#[doc(hidden)]
+pub mod geometry;
+#[doc(hidden)]
+pub mod types;
