+++ crates/auv-driver-macos/src/constants.rs
@@ -1,0 +1,1 @@
+pub const OSASCRIPT_BINARY: &str = "/usr/bin/osascript";
