+++ src/driver/macos/capture/mod.rs
@@ -1,5 +1,6 @@
-// File: src/driver/macos/capture/mod.rs
-pub(crate) mod artifact;
+// TODO(driver-crates): temporary root compatibility while command handlers
+// still produce root `DriverResponse` and runtime artifacts.
+pub(crate) use auv_driver_macos::capture::{artifact, types};
+
 pub(crate) mod commands;
-pub(crate) mod types;
 pub(crate) mod xcap_backend;
