+++ src/driver/macos/constants.rs
@@ -1,2 +1,3 @@
-// File: src/driver/macos/constants.rs
-pub(crate) const OSASCRIPT_BINARY: &str = "/usr/bin/osascript";
+// TODO(driver-crates): temporary root compatibility while legacy command
+// handlers migrate to `auv-driver-macos` typed session APIs.
+pub(crate) use auv_driver_macos::constants::*;
