+++ src/driver/macos/support/geometry.rs
@@ -3,35 +3,10 @@ use super::super::*;
 use super::call::optional_f64;
 
 pub(crate) fn compute_combined_bounds(displays: &[ObservedDisplay]) -> ObservedRect {
-  let min_x = displays
-    .iter()
-    .map(|display| display.bounds.x)
-    .min()
-    .unwrap_or(0);
-  let min_y = displays
-    .iter()
-    .map(|display| display.bounds.y)
-    .min()
-    .unwrap_or(0);
-  let max_x = displays
-    .iter()
-    .map(|display| display.bounds.x + display.bounds.width)
-    .max()
-    .unwrap_or(0);
-  let max_y = displays
-    .iter()
-    .map(|display| display.bounds.y + display.bounds.height)
-    .max()
-    .unwrap_or(0);
-
-  ObservedRect {
-    x: min_x,
-    y: min_y,
-    width: max_x - min_x,
-    height: max_y - min_y,
-  }
+  auv_driver_macos::types::compute_combined_bounds(displays)
 }
 
+#[cfg(test)]
 pub(crate) fn app_contains_window(app_identifier: &str, app_name: &str) -> bool {
   let app_identifier = app_identifier.trim().to_ascii_lowercase();
   let app_name = app_name.trim().to_ascii_lowercase();
@@ -40,15 +15,13 @@ pub(crate) fn app_contains_window(app_identifier: &str, app_name: &str) -> bool 
     || app_name.contains(&app_identifier)
 }
 
+#[cfg(test)]
 pub(crate) fn window_area(window: &ObservedWindow) -> i64 {
   window.bounds.width.saturating_mul(window.bounds.height)
 }
 
 pub(crate) fn ocr_match_center(matched: &OcrTextMatch) -> (f64, f64) {
-  (
-    matched.bounds.x as f64 + (matched.bounds.width as f64 / 2.0),
-    matched.bounds.y as f64 + (matched.bounds.height as f64 / 2.0),
-  )
+  auv_driver_macos::support::ocr_match_center(matched)
 }
 
 pub(crate) fn project_main_screenshot_point(
@@ -56,26 +29,7 @@ pub(crate) fn project_main_screenshot_point(
   screenshot_x: f64,
   screenshot_y: f64,
 ) -> AuvResult<(f64, f64)> {
-  let main_display = snapshot
-    .displays
-    .iter()
-    .find(|display| display.is_main)
-    .or_else(|| snapshot.displays.first())
-    .ok_or_else(|| "display probe returned no connected displays".to_string())?;
-  if screenshot_x < 0.0
-    || screenshot_y < 0.0
-    || screenshot_x >= main_display.pixel_width as f64
-    || screenshot_y >= main_display.pixel_height as f64
-  {
-    return Err(format!(
-      "screenshot pixel point ({screenshot_x:.3}, {screenshot_y:.3}) is outside main display physical bounds {}x{}",
-      main_display.pixel_width, main_display.pixel_height
-    ));
-  }
-  Ok((
-    main_display.bounds.x as f64 + (screenshot_x / main_display.scale_factor),
-    main_display.bounds.y as f64 + (screenshot_y / main_display.scale_factor),
-  ))
+  auv_driver_macos::support::project_main_screenshot_point(snapshot, screenshot_x, screenshot_y)
 }
 
 pub(crate) fn resolve_window_point(
@@ -125,23 +79,7 @@ pub(crate) fn resolve_display_point(
   x: f64,
   y: f64,
 ) -> Option<ObservedPointResolution> {
-  let display = snapshot.displays.iter().find(|display| {
-    let left = display.bounds.x as f64;
-    let top = display.bounds.y as f64;
-    let right = left + display.bounds.width as f64;
-    let bottom = top + display.bounds.height as f64;
-    x >= left && x < right && y >= top && y < bottom
-  })?;
-  let local_x = x - display.bounds.x as f64;
-  let local_y = y - display.bounds.y as f64;
-
-  Some(ObservedPointResolution {
-    display: display.clone(),
-    local_x,
-    local_y,
-    backing_pixel_x: (local_x * display.scale_factor).round() as i64,
-    backing_pixel_y: (local_y * display.scale_factor).round() as i64,
-  })
+  auv_driver_macos::support::resolve_display_point(snapshot, x, y)
 }
 
 pub(crate) fn render_rect_compact(rect: &ObservedRect) -> String {
